Error hash modifications

+++ handlers/trips/delete.py
@@ -1,4 +1,5 @@
 import webapp2
+from google.appengine.api import users
 from google.appengine.ext import ndb
 
 from models.collaboration import Collaboration
@@ -8,17 +9,22 @@ from models.trip import Trip
 class TripDelete(webapp2.RequestHandler):
 
     def post(self):
-        # Get related collaborations
-        collaborations = list(Collaboration.query(Collaboration.trip_key == int(self.request.get("trip_key"))))
+        trip = Trip.get_by_id(int(self.request.get("trip_key")))
 
-        # Delete collaborations
-        for c in collaborations:
-            ndb.Key(Collaboration, int(c.key.id())).delete()
+        if users.get_current_user().email() != trip.owner:
+            self.redirect("/trips/manage?message=e71c6822d05dees5s9a5t76c6e6429af3")
+        else:
+            # Get related collaborations
+            collaborations = list(Collaboration.query(Collaboration.trip_key == int(self.request.get("trip_key"))))
 
-        # Delete wanted trip
-        ndb.Key(Trip, int(self.request.get("trip_key"))).delete()
+            # Delete collaborations
+            for c in collaborations:
+                ndb.Key(Collaboration, int(c.key.id())).delete()
 
-        self.redirect("/trips/manage?message=dc5552978b5eea6cbc607611e7f4025b")
+            # Delete wanted trip
+            ndb.Key(Trip, int(self.request.get("trip_key"))).delete()
+
+            self.redirect("/trips/manage?message=sdc5552978b5eea6cbc607611e7f4025b")
 
 
 app = webapp2.WSGIApplication([('/trips/delete', TripDelete), ], debug=True)
